Fix broken links on the site
We want to support developers who just checkout the dashboard and load
libraries from sleepdiary.github.io

But we also want to support forks that run the dashboard on another
site.

The existing ".env.development" supports the first case, this file
supports the second.  We use this instead of ".env.production" in case
e.g. we need to support a testing build or something in future.